Move ArgValueOfAllowedTypeChecker to top level and cosmetic test cleanup
